+++ src/HttpAdapter/Guzzle5Adapter.php
@@ -3,13 +3,14 @@
 namespace Phpoaipmh\HttpAdapter;
 
 use GuzzleHttp\Client as GuzzleClient;
+use GuzzleHttp\ClientInterface;
 
 /**
  * GuzzleAdapter HttpAdapter HttpAdapterInterface Adapter
  *
  * @package Phpoaipmh\HttpAdapter
  */
-class GuzzleAdapter implements HttpAdapterInterface
+class Guzzle5Adapter implements HttpAdapterInterface
 {
     /**
      * @var array  Default Guzzle Options; these are ignored if GuzzleClient passed to constructor
@@ -17,8 +18,13 @@ class GuzzleAdapter implements HttpAdapterInterface
     private static $defaultGuzzleOptions = [
         'connect_timeout'    => 10,
         'timeout'            => 60,
-        'allow_redirects'    => 3,
         'headers'            => ['User-Agent' => 'PHP OAI-PMH Library'],
+        'allow_redirects'    => [
+            'max'       => 3,
+            'strict'    => true,
+            'referer'   => true,
+            'protocols' => ['http', 'https']
+        ]
     ];
 
     /**
@@ -30,10 +36,22 @@ class GuzzleAdapter implements HttpAdapterInterface
      * Constructor
      *
      * @param GuzzleClient $guzzle
+     * @throws \Exception  If using wrong Guzzle version
      */
     public function __construct(GuzzleClient $guzzle = null)
     {
-        $this->guzzle = $guzzle ?: new GuzzleClient(static::$defaultGuzzleOptions);
+        if (substr(ClientInterface::VERSION, 0, 1) != '5') {
+            throw new \Exception(sprintf(
+                'Guzzle is at Version %s.  %s requires Guzzle v5 (should you be using %s instead?)',
+                ClientInterface::VERSION,
+                get_called_class(),
+                GuzzleAdapter::class
+            ));
+        }
+
+        $this->guzzle = $guzzle ?: new GuzzleClient([
+            'defaults' => self::$defaultGuzzleOptions
+        ]);
     }
 
     /**
@@ -49,13 +67,13 @@ class GuzzleAdapter implements HttpAdapterInterface
     /**
      * Do the request with GuzzleAdapter
      *
-     * @param  string  $url
-     * @param  array   $queryParams
+     * @param string $url
+     * @param array  $queryParams
      * @return string
      */
     public function request($url, array $queryParams = [])
     {
-        $response = $this->guzzle->get($url, ['query' => $queryParams]);
-        return (string) $response->getBody();
+        $resp = $this->guzzle->get($url, ['query' => $queryParams]);
+        return (string) $resp->getBody();
     }
 }